+++ src/main/kotlin/io/unthrottled/theme/randomizer/services/AppService.kt
@@ -1,0 +1,8 @@
+package io.unthrottled.theme.randomizer.services
+
+import com.intellij.openapi.application.ApplicationNamesInfo
+
+object AppService {
+  fun getApplicationName(): String =
+    ApplicationNamesInfo.getInstance().fullProductNameWithEdition
+}