+++ src/main/kotlin/io/unthrottled/theme/randomizer/themes/ThemeSelectionService.kt
@@ -1,0 +1,53 @@
+package io.unthrottled.theme.randomizer.themes
+
+import com.intellij.openapi.application.ApplicationManager
+import io.unthrottled.theme.randomizer.config.Config
+
+class ThemeSelectionService {
+  companion object {
+    val instance: ThemeSelectionService
+      get() = ApplicationManager.getApplication().getService(ThemeSelectionService::class.java)
+  }
+
+  private var _preferredThemeIds: Set<String>
+
+  private var _blackListedThemeIds: Set<String>
+
+  init {
+    val currentThemeSelections = getThemeSelectionsFromConfigs()
+    _preferredThemeIds = currentThemeSelections.preferredThemeIdSet
+    _blackListedThemeIds = currentThemeSelections.blacklistedThemeIdSet
+  }
+
+  private fun getThemeSelectionsFromConfigs() =
+    ThemeSelectionExtractor.extractThemeSelectionsFromConfig(Config.instance)
+
+  fun rehydrateSelectionsFromConfig(config: Config) {
+    val currentThemeSelections = ThemeSelectionExtractor.extractThemeSelectionsFromConfig(config)
+    rehydrateFromSelections(currentThemeSelections)
+  }
+
+  private fun rehydrateFromSelections(currentThemeSelections: ThemeSelections) {
+    _preferredThemeIds = currentThemeSelections.preferredThemeIdSet
+    _blackListedThemeIds = currentThemeSelections.blacklistedThemeIdSet
+  }
+
+  fun reHydrateSelections(isLocalSync: Boolean = Config.instance.isLocalSync) {
+    // todo: might need to not run on AWT thread.
+    if (isLocalSync) {
+      rehydrateFromSelections(LocalThemeSelectionService.getInitialItem())
+    } else {
+      rehydrateFromSelections(getThemeSelectionsFromConfigs())
+    }
+  }
+
+  val preferredThemeIds: Set<String>
+    get() {
+      return _preferredThemeIds
+    }
+
+  val blackListedThemeIds: Set<String>
+    get() {
+      return _blackListedThemeIds
+    }
+}