+++ docs/RELEASE-NOTES.md
@@ -2,3 +2,4 @@
 
 - Match OS mode, which allows the plugin to randomly change your theme when it observes your system settings change. [See docs for more details](https://github.com/Unthrottled/theme-randomizer/README.md#settings)
 - 2022.2 Build Support
+- Local theme selection settings synchronization.